+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * don't complete default commands as argument [#465](https://github.com/jcubic/jquery.terminal/issues/465)
 * don't show links in echo command + format links inside formatting [#464](https://github.com/jcubic/jquery.terminal/issues/464)
 * fix calling set_mask in onPop
+* fix visibility change when terminal have `position:fixed` [#466](https://github.com/jcubic/jquery.terminal/issues/466)
 
 ## 2.1.2
 ### Bugfix